highlight module list items

+++ src/components/TopicPillItem.js
@@ -5,13 +5,23 @@ class TopicPillItem extends React.Component{
     constructor(props){
         super(props);
     }
+
+    deleteTopic() {
+        if (window.confirm("Are you sure you want to delete this topic?")) {
+            this.props.deleteTopic(this.props.topic.id);
+        }
+        else {
+            return;
+        }
+    }
+
     render(){
         return(
             <li className="nav-item">
                 <div className="nav-link">
                     {this.props.topic.title}
                     <i style={{'marginLeft': '10px'}} onClick={() => {
-                        this.props.deleteTopic(this.props.topic.id)
+                        this.deleteTopic()
                         // console.log('hello: '+this.props.moduleId)
                         }
                     } className="fa fa-times"></i>
